+++ packages/dataviews/src/dataviews-layouts/picker-grid/index.tsx
@@ -306,9 +306,8 @@ function ViewPickerGrid< Item >( {
 	 */
 	const size = '900px';
 
-	const groupFieldId = view.groupBy?.field || view.groupByField;
-	const groupField = groupFieldId
-		? fields.find( ( f ) => f.id === groupFieldId )
+	const groupField = view.groupBy?.field
+		? fields.find( ( f ) => f.id === view.groupBy?.field )
 		: null;
 	const dataByGroup = groupField ? getDataByGroup( data, groupField ) : null;